restructure code and vote field
