+++ src/Components/GetId.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {useParams} from "react-router-dom";
-import RestaurantInfoPage from './RestaurantInfoPage';
+import RestaurantInfoPage from './Restaurant/RestaurantInfoPage';
 
 export default function GetId() {
