+++ src/Components/Restaurant/RestaurantCardBack.js
@@ -2,8 +2,8 @@ import React from 'react'
 import {NavLink, Redirect} from 'react-router-dom'
 import { connect } from 'react-redux'
 import {Card, Button} from 'semantic-ui-react'
-import { addRestaurantToProfile, removeFromProfile, updateStats } from '../Actions/UserActions'
-import VoteField from './VoteField'
+import { addRestaurantToProfile, removeFromProfile, updateStats } from '../../Actions/UserActions'
+import VoteFieldContainer from '../../Containers/VoteFieldContainer'
 
 
 class RestaurantCardBack extends React.Component {
@@ -20,13 +20,6 @@ class RestaurantCardBack extends React.Component {
         this.props.flipCard()
         this.props.removeFromProfile(this.props.stats.find(rel => rel.restaurant_id === this.props.restaurant.id), this.props.restaurant)
     }
-
-    // redirectToProfile = () => {
-    //     this.props.updateNavbar("profile")
-    //         this.setState({
-    //             redirect: true
-    //         })
-    // }
 
     handleDoneSubmit = (event) => {
         event.preventDefault()
@@ -58,7 +51,7 @@ class RestaurantCardBack extends React.Component {
                 <Card.Description>{this.props.restaurant.city}, {this.props.restaurant.state} {this.props.restaurant.zip_code}</Card.Description>
                 <br></br>
                 <Card.Header>Brunchr Highlights</Card.Header>
-                <VoteField restaurant={this.props.restaurant}></VoteField>
+                <VoteFieldContainer restaurant={this.props.restaurant}></VoteFieldContainer>
                 <br></br>
                 {profileButton()}<NavLink to={link}><Button size="small">Info</Button></NavLink><Button size="small" onClick={this.handleDoneSubmit}>Done</Button>
                 </>
